backend/requirement.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.14.0
marshmallow==3.20.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==42.0.5
